Split up engine, power events and power iteration into modules
